pyproject.toml
[project]
name = "openpaw"
version = "0.1.0"
description = "OpenPaw - Multi-Channel AI Agent Framework with LangGraph"
authors = [
    {name = "John Sosoka"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=1.2.8,<2.0.0)",
    "langchain-anthropic (>=1.3.1,<2.0.0)",
    "langchain-aws (>=1.0.0,<2.0.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "langgraph (>=1.0.7,<2.0.0)",
    "langgraph-checkpoint-sqlite (>=2.0.0,<4.0.0)",
    "python-telegram-bot[job-queue] (>=22.6,<23.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "apscheduler (>=3.10.0,<4.0.0)",
    "langchain-experimental (>=0.4.1,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "docling (>=2.72.0,<3.0.0)",
    "playwright (>=1.58.0,<2.0.0)"
]

[project.optional-dependencies]
# Builtins - install specific extras for the builtins you need
voice = [
    "openai (>=1.0.0)",      # Whisper transcription
    "elevenlabs (>=1.0.0)"   # Text-to-speech
]
web = [
    "langchain-community (>=0.4.0)"  # Brave Search
]
system = [
    "langchain-experimental (>=0.4.0)",  # Shell tool
    "asyncssh (>=2.0.0)"  # SSH remote execution
]
all-builtins = [
    "openai (>=1.0.0)",
    "elevenlabs (>=1.0.0)",
    "langchain-community (>=0.4.0)",
    "langchain-experimental (>=0.4.0)",
    "asyncssh (>=2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
system = [
    "asyncssh (>=2.22.0,<3.0.0)"
]

[project.scripts]
openpaw = "openpaw.cli:run"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Deprecation shims: Allow imports after warnings
"openpaw/commands/*.py" = ["E402"]
"openpaw/commands/handlers/*.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
